MOO-150: separated auditService
