+++ api/src/main/java/org/openmrs/module/msfcore/aop/ViewPatientAdvice.java
@@ -5,7 +5,7 @@ import java.lang.reflect.Method;
 import org.openmrs.Patient;
 import org.openmrs.User;
 import org.openmrs.api.context.Context;
-import org.openmrs.module.msfcore.api.MSFCoreService;
+import org.openmrs.module.msfcore.api.AuditService;
 import org.openmrs.module.msfcore.audit.MSFCoreLog;
 import org.openmrs.module.msfcore.audit.MSFCoreLog.Event;
 import org.springframework.aop.AfterReturningAdvice;
@@ -26,7 +26,7 @@ public class ViewPatientAdvice implements AfterReturningAdvice {
           "loaded/viewed patient#" + patient.getPatientIdentifier().getIdentifier(), Context.getAuthenticatedUser());
       viewPatientLog.setPatient(patient);
       viewPatientLog.setUser((User) args[1]);
-      Context.getService(MSFCoreService.class).saveMSFCoreLog(viewPatientLog);
+      Context.getService(AuditService.class).saveMSFCoreLog(viewPatientLog);
     }
   }
