+++ api/src/main/java/org/openmrs/module/msfcore/api/impl/MSFCoreServiceImpl.java
@@ -10,17 +10,12 @@
 package org.openmrs.module.msfcore.api.impl;
 
 import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Date;
 import java.util.List;
 
 import org.openmrs.Concept;
 import org.openmrs.Location;
 import org.openmrs.LocationAttribute;
 import org.openmrs.LocationAttributeType;
-import org.openmrs.Patient;
-import org.openmrs.Provider;
-import org.openmrs.User;
 import org.openmrs.api.APIException;
 import org.openmrs.api.context.Context;
 import org.openmrs.api.impl.BaseOpenmrsService;
@@ -29,8 +24,6 @@ import org.openmrs.module.idgen.SequentialIdentifierGenerator;
 import org.openmrs.module.msfcore.DropDownFieldOption;
 import org.openmrs.module.msfcore.api.MSFCoreService;
 import org.openmrs.module.msfcore.api.dao.MSFCoreDao;
-import org.openmrs.module.msfcore.audit.MSFCoreLog;
-import org.openmrs.module.msfcore.audit.MSFCoreLog.Event;
 
 public class MSFCoreServiceImpl extends BaseOpenmrsService implements MSFCoreService {
 
@@ -41,25 +34,6 @@ public class MSFCoreServiceImpl extends BaseOpenmrsService implements MSFCoreSer
    */
   public void setDao(MSFCoreDao dao) {
     this.dao = dao;
-  }
-
-  public List<MSFCoreLog> getMSFCoreLogs(Date startDate, Date endDate, List<Event> events, User creator, List<Patient> patients,
-          List<User> users, List<Provider> providers, List<Location> locations) {
-    return dao.getMSFCoreLogs(startDate, endDate, events, creator, patients, users, providers, locations);
-  }
-
-  public MSFCoreLog getMSFCoreLogByUuid(String uuid) {
-    return dao.getMSFCoreLogByUuid(uuid);
-  }
-
-  public void deleteMSFCoreLog(MSFCoreLog msfCoreLog) {
-    dao.deleteMSFCoreLog(msfCoreLog);
-  }
-
-  public void deleteMSFCoreFromDate(Date startDate) {
-    for (MSFCoreLog log : getMSFCoreLogs(startDate, null, null, null, null, null, null, null)) {
-      deleteMSFCoreLog(log);
-    }
   }
 
   public List<Concept> getAllConceptAnswers(Concept question) {
@@ -80,20 +54,5 @@ public class MSFCoreServiceImpl extends BaseOpenmrsService implements MSFCoreSer
       answerNames.add(new DropDownFieldOption(String.valueOf(answer.getId()), answer.getName().getName()));
     }
     return answerNames;
-  }
-
-  public Date getDateAtNDaysFromData(Date date, Integer nDays) {
-    Calendar calendar = Calendar.getInstance();
-    calendar.setTime(date);
-    calendar.add(Calendar.DAY_OF_YEAR, -nDays);
-    return calendar.getTime();
-  }
-
-  public Integer saveMSFCoreLog(MSFCoreLog msfCoreLog) {
-    return dao.saveMSFCoreLog(msfCoreLog);
-  }
-
-  public MSFCoreLog getMSFCoreLog(Integer msfCoreLogId) {
-    return dao.getMSFCoreLog(msfCoreLogId);
   }
 }
